+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -5827,7 +5827,6 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return handler(register_name)
 
     class WriteRegisterCommand(FastCommand):
-        # pylint: disable=line-too-long
         """
         Class for handling the WriteRegister() command.
 
@@ -6233,7 +6232,6 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return handler(argin)
 
     class SetLmcDownloadCommand(FastCommand):
-        # pylint: disable=line-too-long
         """
         Class for handling the SetLmcDownload() command.
 
@@ -6446,7 +6444,6 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return ([return_code], [message])
 
     class SetCspDownloadCommand(FastCommand):
-        # pylint: disable=line-too-long
         """
         Class for handling the SetCspDownload() command.
 
@@ -7746,7 +7743,6 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return ([return_code], [message])
 
     class SendDataSamplesCommand(FastCommand):
-        # pylint: disable=line-too-long
         """
         Class for handling the SendDataSamples() command.